Create copy of oem151 with empty lists
A copy of oem151 is created (v151_temp_backfill_clean), in which the fields that contain template dicts are emptied. This serves the purpose of adding only the amount of dicts that are actually in oem141. In doing so, the template dicts in oem151 don't need to be deleted later on to get an exact copy of the in information in oem141.
#50