Delete code draft for detailed copy of foreignKeys
In the current implementation the 'foreignKeys' are copied as a complete copy the list in oem141. This means if sub-fields in 'foreignKeys' are missing they will not be updated with default values of oem151. The reason is that currently some 'foreignKeys' default values cause errors when uploading to the OEP. This implementation is chosen under the assumption that metadata in oem141 on the OEP has already gone through a validity check. Thus, the 'foreignKeys' should be valid.
#50